feat: useStepManager 훅 추가 및 AddCardPage에서 상태 관리 로직 통합
